start implementing some routes and move some files around
