+++ client/src/common/components/index.js
@@ -1,0 +1,3 @@
+export { default as Container } from "./Container";
+export { default as FluidContainer } from "./FluidContainer";
+export { default as Navbar } from "./Navbar";
