+++ client/src/pages/Dashboard/components/Account.js
@@ -1,5 +1,6 @@
 import React from "react";
 
+// TODO: Consider displaying accounts as a grid of cards on the dashboard.
 const Account = ({ id, name, type }) => (
     <li
         id={id}
