+++ client/src/pages/Dashboard/components/index.js
@@ -1,0 +1,2 @@
+export { default as Account } from "./Account";
+export { default as AccountList } from "./AccountList";
