+++ client/src/pages/Dashboard/index.js
@@ -1,0 +1,3 @@
+import DashboardPage from "./DashboardPage"
+
+export { DashboardPage as Dashboard };
