+++ client/src/pages/Landing/LandingPage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function LandingPage() {
+    return (
+        <h1>Landing Page</h1>
+    )
+}
+
+export default LandingPage;
