+++ client/src/pages/Landing/index.js
@@ -1,0 +1,3 @@
+import LandingPage from "./LandingPage"
+
+export { LandingPage as Landing };
